remove-image-bg/requirements.txt
pil2ansi==1.0.0
rembg==2.0.62
onnxruntime==1.20.1